requirements.txt
aniso8601==9.0.1
astroid==2.11.7
black==22.6.0
blinker==1.5
click==8.1.3
dill==0.3.5.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pylint==2.14.5
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
